+++ src/nplinker/scoring/linking/__init__.py
@@ -1,6 +1,6 @@
 import logging
-from .data_linking import DataLinks
-from .data_linking import LINK_TYPES
+from .data_links import DataLinks
+from .data_links import LINK_TYPES
 from .data_linking_functions import calc_correlation_matrix
 from .link_finder import LinkFinder
